+++ Middleware/SymfonyRouteResolver.php
@@ -38,10 +38,10 @@ use Symfony\Component\Routing\Exception\MethodNotAllowedException;
 use Symfony\Component\Routing\Exception\ResourceNotFoundException;
 use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Context\LanguageAspectFactory;
-use TYPO3\CMS\Core\Routing\RouteNotFoundException;
-use TYPO3\CMS\Core\Routing\SiteRouteResult;
+use TYPO3\CMS\Core\Routing\PageArguments;
 use TYPO3\CMS\Core\Site\Entity\Site;
 use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
+use TYPO3\CMS\Frontend\Middleware\PrepareTypoScriptFrontendRendering;
 use TYPO3\CMS\Frontend\Middleware\TypoScriptFrontendInitialization;
 use TYPO3\CMS\Frontend\Page\PageInformationCreationFailedException;
 
@@ -56,6 +56,7 @@ class SymfonyRouteResolver implements MiddlewareInterface
         private readonly Context $context,
         private readonly RequestHandlerInterface $dummyRequestHandler,
         private readonly TypoScriptFrontendInitialization $frontendInitialization,
+        private readonly PrepareTypoScriptFrontendRendering $typoScriptFrontendRendering,
     ) {
         $this->httpFoundationFactory = new HttpFoundationFactory();
 
@@ -85,32 +86,21 @@ class SymfonyRouteResolver implements MiddlewareInterface
 
     private function initializeTemporaryTSFE(ServerRequestInterface $request): void
     {
+        // do nothing if TSFE is already initialized
         if ($GLOBALS['TSFE'] instanceof TypoScriptFrontendController) {
             return;
         }
 
+        // abort if site not found
         $site = $request->getAttribute('site');
         if (!$site instanceof Site) {
             return;
         }
 
+        // simulate a TYPO3 request to the site's root page in order to generate the TSFE + template config, ...
+        $fakePageArguments = new PageArguments($site->getRootPageId(), '0', []);
         $previousRouting = $request->getAttribute('routing');
-        if (!$previousRouting instanceof SiteRouteResult) {
-            return;
-        }
-
-        $fakeRouting = new SiteRouteResult(
-            $previousRouting->getUri(),
-            $previousRouting->getSite(),
-                $previousRouting->getLanguage() ?? $site->getDefaultLanguage()
-        );
-
-        try {
-            $pageArguments = $site->getRouter()->matchRequest($request, $fakeRouting);
-            $request = $request->withAttribute('routing', $pageArguments);
-        } catch (RouteNotFoundException) {
-            // route cannot be found if it's a symfony route
-        }
+        $request = $request->withAttribute('routing', $fakePageArguments);
 
         // set language aspect
         $language = $request->getAttribute('language', $site->getDefaultLanguage());
@@ -118,11 +108,20 @@ class SymfonyRouteResolver implements MiddlewareInterface
         $this->context->setAspect('language', $languageAspect);
 
         try {
+            // call TYPO3 middlewares to generate the TSFE + TypoScript config
+            // and keep the $request up-to-date with all new attributes
             $this->frontendInitialization->process($request, $this->dummyRequestHandler);
+            /** @noinspection CallableParameterUseCaseInTypeContextInspection */
+            $request = $GLOBALS['TYPO3_REQUEST'];
+
+            $this->typoScriptFrontendRendering->process($request, $this->dummyRequestHandler);
+            /** @noinspection CallableParameterUseCaseInTypeContextInspection */
+            /** @noinspection PhpConditionAlreadyCheckedInspection */
+            $request = $GLOBALS['TYPO3_REQUEST'];
         } catch (PageInformationCreationFailedException) {
         }
 
-        // unset the changes
+        // unset the routing changes
         $request = $request->withAttribute('routing', $previousRouting);
         $GLOBALS['TYPO3_REQUEST'] = $request;
     }