+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Set minimum Symfony version to 7.1
 - Set minimum php version to 8.3
 
+### Changed
+- Creating TSFE and template config in Symfony context
+
 ### Removed
 - TYPO3 helhum console
-- TSFE subscriber to load TYPO3 stuff for Symfony routes
 - Converting TYPO3 request to PSR request in `$GLOBALS['TYPO3_REQUEST']`
 
 ## [5.1.0] - 2024-10-04